Created a point light shader to physically display a circle where the point light supposed to be

+++ src/descriptors.cpp
@@ -205,7 +205,7 @@ namespace grape {
         for (auto& write : writes) {
             write.dstSet = set;
         }
-        vkUpdateDescriptorSets(pool.grapeDevice.device(), writes.size(), writes.data(), 0, nullptr);
+        vkUpdateDescriptorSets(pool.grapeDevice.device(), static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
     }
 
 }
