+++ src/renderer.cpp
@@ -130,7 +130,7 @@ namespace grape {
 		vkFreeCommandBuffers(
 			grapeDevice.device(),
 			grapeDevice.getCommandPool(),
-			static_cast<float>(commandBuffers.size()),
+			static_cast<uint32_t>(commandBuffers.size()),
 			commandBuffers.data());
 		commandBuffers.clear();
 	}
